hcm config: Split FilterChainTest into separate file (#15048)
Split test functions into config_test_base.h
Move FilterChainTest tests into config_filter_chain_test.cc
Other tests from config_test.cc can be moved in the future to improve readability

Risk Level: Low
As part of #14470, I'll need to add filter dependency validation logic for http filterchains. This is a cleanup CL to make that easier.
